fix : refactoring -1
