--- YogiYoCloneIOS/YogiYoCloneIOS/ViewController/ViewControllers/HomeVC/Byun/DetailVC/NilCell.swift
@@ -1,24 +1,0 @@
-//
-//  NilCell.swift
-//  YogiYoCloneIOS
-//
-//  Created by Qussk_MAC on 2020/08/25.
-//  Copyright © 2020 김동현. All rights reserved.
-//
-
-import UIKit
-
-class NilCell: UITableViewCell {
-
-    override func awakeFromNib() {
-        super.awakeFromNib()
-        // Initialization code
-    }
-
-    override func setSelected(_ selected: Bool, animated: Bool) {
-        super.setSelected(selected, animated: animated)
-
-        // Configure the view for the selected state
-    }
-
-}
