--- YogiYoCloneIOS/YogiYoCloneIOS/ViewController/ViewControllers/HomeVC/Byun/DetailVC/View/CancelBt.swift
@@ -1,14 +1,0 @@
-//
-//  CancelBt.swift
-//  YogiYoCloneIOS
-//
-//  Created by Qussk_MAC on 2020/08/26.
-//  Copyright © 2020 김동현. All rights reserved.
-//
-
-import UIKit
-
-class CancelBt : UIButton{
-  
-  
-}
