+++ YogiYoCloneIOS/YogiYoCloneIOS/ViewController/ViewControllers/HomeVC/Byun/OrderVC/Controller/OrderVC.swift
@@ -28,7 +28,6 @@ class OderVC : UIViewController {
     super.viewDidLoad()
     orderList = orderMager.showMeOrderedList()
     setTableView2()
-    buttonFrame()
     //  navigation()
     title = "배달 주문 결제"
     //   fetchPosts()
@@ -47,13 +46,11 @@ class OderVC : UIViewController {
   
   //MARK:- POST
   func onPostShowBible(){
-    print("포스트 방식 데이터 가지러옴")
     let datalist = orderList
     guard let url = URL(string: "http://52.79.251.125/orders") else {return}
     var request = URLRequest(url: url)
     request.httpMethod = "POST"
-    // 4. HTTP 메시지에 포함될 헤더 설정
-    request.addValue("application/json", forHTTPHeaderField: "text/html")
+    request.addValue("application/json", forHTTPHeaderField: "Content-Type")
     let body = "http://52.79.251.125/orders".data(using:String.Encoding.ascii, allowLossyConversion: false)
     
     request.httpBody = body
@@ -173,78 +170,76 @@ class OderVC : UIViewController {
     self.present(alert, animated: true, completion: nil)
   }
 }
-
-
-//MARK:-UITableViewDataSource
-extension OderVC : UITableViewDataSource{
-  func numberOfSections(in tableView: UITableView) -> Int {
-    return 6
-  }
-  
-  func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-    switch section {
-    case 0:  //로그인 유무
-      return 1
-    case 1://주문자정보
-      if open == false {
-        return 2
-      }else {
-        return 2 + 1
-      }
-    case 2: //결제수단 선택
-      return 1
-    case 3: //할인방법 선택
-      return 1
-    case 4: //배달주문 내역
-      return 1
-    default:
-      return 1
-    }
-  }
-  
-  func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-    switch indexPath.section {
-    case 0:
-      let loginCell = tableView.dequeueReusableCell(withIdentifier: "loginCell", for: indexPath) as! loginCell
-      return loginCell
-    case 1:
-      switch indexPath.row {
-      case 0:
-        let InformationCell = tableView.dequeueReusableCell(withIdentifier: "InformationCell", for: indexPath) as! InformationCell
-        return InformationCell
-      case 1:
-        let CustomOderCell = tableView.dequeueReusableCell(withIdentifier: "CustomOrderCell", for: indexPath) as! CustomOrderCell
-        _ = false
-        CustomOderCell.configure(title: "\(userString)")
-        // print(userString)
-        return CustomOderCell
-        
-      default:
-        let unMembershipCell = tableView.dequeueReusableCell(withIdentifier: "unMembershipCell", for: indexPath) as!
-          unMembershipCell
-        return unMembershipCell
-      }
-    case 2:
-      let PaywithCell = tableView.dequeueReusableCell(withIdentifier: "PaywithCell", for: indexPath) as! PaywithCell
-      return PaywithCell
-    case 3:
-      let MembershipCell = tableView.dequeueReusableCell(withIdentifier: "MembershipCell", for: indexPath) as! MembershipCell
-      return MembershipCell
-    case 4:
-      let OrderListCell = tableView.dequeueReusableCell(withIdentifier: "OrderListCell", for: indexPath) as! OrderListCell
-      print("오더리스트셀 제발 알려주세요", indexPath)
-      OrderListCell.orderData = orderList
-      
-      return OrderListCell
-    case 5 :
-      let paymentCell = tableView.dequeueReusableCell(withIdentifier: "paymentCell", for: indexPath) as! paymentCell
-      return paymentCell
-    default:
-      let loginCell = tableView.dequeueReusableCell(withIdentifier: "loginCell", for: indexPath) as! loginCell
-      return loginCell
-    }
-  }
-}
+////MARK:-UITableViewDataSource
+//extension OderVC : UITableViewDataSource{
+//  func numberOfSections(in tableView: UITableView) -> Int {
+//    return 6
+//  }
+//  
+//  func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+//    switch section {
+//    case 0:  //로그인 유무
+//      return 1
+//    case 1://주문자정보
+//      if open == false {
+//        return 2
+//      }else {
+//        return 2 + 1
+//      }
+//    case 2: //결제수단 선택
+//      return 1
+//    case 3: //할인방법 선택
+//      return 1
+//    case 4: //배달주문 내역
+//      return 1
+//    default:
+//      return 1
+//    }
+//  }
+//  
+//  func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+//    switch indexPath.section {
+//    case 0:
+//      let loginCell = tableView.dequeueReusableCell(withIdentifier: "loginCell", for: indexPath) as! loginCell
+//      return loginCell
+//    case 1:
+//      switch indexPath.row {
+//      case 0:
+//        let InformationCell = tableView.dequeueReusableCell(withIdentifier: "InformationCell", for: indexPath) as! InformationCell
+//        return InformationCell
+//      case 1:
+//        let CustomOderCell = tableView.dequeueReusableCell(withIdentifier: "CustomOrderCell", for: indexPath) as! CustomOrderCell
+//        _ = false
+//        CustomOderCell.configure(title: "\(userString)")
+//        // print(userString)
+//        return CustomOderCell
+//        
+//      default:
+//        let unMembershipCell = tableView.dequeueReusableCell(withIdentifier: "unMembershipCell", for: indexPath) as!
+//          unMembershipCell
+//        return unMembershipCell
+//      }
+//    case 2:
+//      let PaywithCell = tableView.dequeueReusableCell(withIdentifier: "PaywithCell", for: indexPath) as! PaywithCell
+//      return PaywithCell
+//    case 3:
+//      let MembershipCell = tableView.dequeueReusableCell(withIdentifier: "MembershipCell", for: indexPath) as! MembershipCell
+//      return MembershipCell
+//    case 4:
+//      let OrderListCell = tableView.dequeueReusableCell(withIdentifier: "OrderListCell", for: indexPath) as! OrderListCell
+//      print("오더리스트셀 제발 알려주세요", indexPath)
+//      OrderListCell.orderData = orderList
+//      
+//      return OrderListCell
+//    case 5 :
+//      let paymentCell = tableView.dequeueReusableCell(withIdentifier: "paymentCell", for: indexPath) as! paymentCell
+//      return paymentCell
+//    default:
+//      let loginCell = tableView.dequeueReusableCell(withIdentifier: "loginCell", for: indexPath) as! loginCell
+//      return loginCell
+//    }
+//  }
+//}
 //MARK:-UITableViewDelegate
 extension OderVC : UITableViewDelegate {
   func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
